requirements.txt
djangorestframework>=3.8.0

